feat: adds hot reload (#53)
* adds hot reload

* updates yarn.lock
